busy with front end

+++ FlintstonesFeedWorkerService/FeedModel.cs
@@ -12,7 +12,7 @@ namespace FlintstonesFeedWorkerService
     {
         public string PartitionKey { get; set; }
         public string RowKey { get; set; }
-        public double LastPrice { get; set; }
+        public decimal LastPrice { get; set; }
         public DateTimeOffset? Timestamp { get ; set ; }
         public ETag ETag { get; set; }
     }
